专利ID类型改为long

+++ src/main/java/com/lotut/pms/dao/PatentDao.java
@@ -56,6 +56,6 @@ public interface PatentDao {
 	
 	Patent getPatentsByAppNo(Integer userId,String appNo);
 	
-	int getPatentIdByAppNo(Integer userId,String appNo);
+	long getPatentIdByAppNo(Integer userId,String appNo);
 	
 }
