+++ src/main/java/com/lotut/pms/service/PatentService.java
@@ -56,7 +56,7 @@ public interface PatentService {
 	 
 	 Patent getPatentsByAppNo(Integer userId,String appNo);
 	 
-	 int getPatentIdByAppNo(Integer userId,String appNo);
+	 long getPatentIdByAppNo(Integer userId,String appNo);
 	 
 	 void autoUpdatePatents(InputStream is,int userId) throws IOException;
